Improved grammar/typos/readability & formatting
There were a few typos in the page, as well as grammar errors.  Changed
the headers in the troubleshoot section as bold, and broke up
wrap-around text in the answers section of troubleshooting.  Improved
readability in some text sections.  Properly
formatted the wallet-address code section to be more easily readable
with some extra css is dd17fe9825022be7.css